Added win condition, all basic objectives now met
UI now updates and player movement disabled when player has collected all items and enters the finish zone. Added a game state system to the GameManager to handle this. The check for whether the player has won is handled in the WinBoxCollisionHandler.cs class, which is assigned to an invisible trigger box in the finish zone.

Additionally added some level boundaries and renamed the CollisionHandler.cs class to StarCollisionHandler.cs, as CollisionHandler was too generic and there are now other game objects besides the star pickups requiring their own collision handling scripts.