Added art assets to player
Added new mesh and animations for player using assets from the Unity asset store. Updated the player controller so that the player now rotates based on movement input, and added handling for updating animation states.

Also updated some handling of unsubscribing to delegates, as Unity was throwing errors for not being able to find instances when attempting to unsubscribe on exit.