Added dog character, audio manager
Dog character responds to player hits. This is handled in the PlayerController class, and uses a delegate to notify the DogManager to respond to being hit.

With more sound effects being added I added an AudioManager. This contains an array of type custom class Sound, and allows other scripts to play sound effects by passing in a sound effect name string, rather than having to rely on array index integers which can be easy to make mistakes with.

Manager instances are now assigned to their own game objects which exist as children under the GameManager object. How it was setup previously, where managers were components assigned to the GameManager object, did not allow for the desired functionality for the AudioManager and I thought it was best to make all managers consistent.