Dog now speaks
Added Text3D.cs class for handling of displaying 3D text. Used this so the dog character can now have various responses depending on if the player has or has not yet collected all the stars, and if the player hits them.

As the DogController now handles the win condition and updating the game state, I deleted the previous collision box and collision handler script.

Also, camera now follows player.